items/requirements.txt
flask
opentelemetry-instrumentation
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
mysql-connector-python  # Only in items service
opentelemetry-instrumentation-mysql  # Only in items service
